Added set-up screen, reorganised widget - New set-up screen, it now properly handles if timetables are nil - Added better organisation for files - Literally go look at the file directory
