+++ Timetables/Lesson/WidgetLesson.swift
@@ -1,0 +1,15 @@
+//
+//  LessonStruct.swift
+//  Timetables
+//
+//  Created by JiaChen(: on 6/7/20.
+//  Copyright © 2020 SST Inc. All rights reserved.
+//
+
+import Foundation
+
+public struct WidgetLesson {
+    public var name: String
+    public var date: Date
+    public var imageName: String
+}
